corto/pyproject.toml
[project]
name = "corto"
version = "0.1.0"
description = ""
authors = [{ name = "Shridhar Shah", email = "shridhsh@cisco.com" }]
requires-python = ">=3.13,<4.0"
dependencies = [
    "a2a-sdk==0.3.0",
    "click>=8.1.8",
    "coloredlogs>=15.0.1",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "langchain-anthropic>=0.3.13",
    "langchain-google-genai>=2.1.4",
    "langchain-openai>=0.3.16",
    "langchain-litellm>=0.3.0",
    "langgraph>=0.4.1",
    "langgraph-supervisor>=0.0.26",
    "litellm==1.75.3",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "requests",
    "starlette",
    "uvicorn",
    "ioa-observe-sdk==1.0.24",
    "agntcy-app-sdk==0.4.5",
    "litellm[proxy]==1.75.3",
    "azure-core==1.38.0",

    # Vulnerability fix for requests
    "langchain-core>=1.2.5",
    "urllib3>=2.6.3",
    "aiohttp>=3.13.3",
    "pyasn1>=0.6.2",
    "pypdf>=6.6.0",
    "filelock>=3.20.3",
    "marshmallow>=3.26.2",
    "PyNaCl>=1.6.2",
]

[project.optional-dependencies]
dev = [
    "pytest~=7.0",
    "pytest-asyncio>=0.17.0,<0.18",
    "pytest-cov>=4.0.0,<5",
    "typing-extensions>=4.12.2,<5",
    "openai>=1.86.0,<2.0",
    "autogen-core>=0.4.3,<0.5",
    "coloredlogs>=15.0.1,<16",
    "langchain-openai>=0.3.14,<0.4",
    "sentence-transformers>=5.1.1",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aml314b", "common", "config", "fi_a", "fi_b"]
